refactor: break classes and enums into separate files
