+++ Dalamud/Game/Internal/Gui/Toast/QuestToastPosition.cs
@@ -1,0 +1,9 @@
+﻿namespace Dalamud.Game.Internal.Gui.Toast
+{
+    public enum QuestToastPosition
+    {
+        Centre = 0,
+        Right = 1,
+        Left = 2,
+    }
+}
