+++ Dalamud/Game/Internal/Gui/Toast/ToastPosition.cs
@@ -1,0 +1,8 @@
+﻿namespace Dalamud.Game.Internal.Gui.Toast
+{
+    public enum ToastPosition : byte
+    {
+        Bottom = 0,
+        Top = 1,
+    }
+}
